mac only code
This code will tell your mac speak "Goodbye cruel world" and prompt for turning off the computer. I would like to implement the prompt to a dialog box input for turning off the computer as a new pop up window outside the prompt so I can apply it as a hotkey.